Take the content blocker and 13 more files off dart:io
The adblock engine gets a seam: dart:ffi on native, a half on web whose
constructors return null exactly as the native half does when the shared
library is missing, so callers keep their existing "engine unavailable" path.
FileStore grows byte support for the serialized engine cache. Content blocker
and DNS blocklist move their caches onto it, and every file whose only dart:io
use was Platform moves to the host shim, including settings, app_settings and
inappbrowser.
