+++ lib/platform/host_platform_web.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 // Web has no host platform in the dart:io sense: nothing here is Android, iOS
 // or macOS, so every per-OS branch takes its non-native path.
 bool get hostIsAndroid => false;
@@ -6,6 +8,7 @@ bool get hostIsMacOS => false;
 bool get hostIsLinux => false;
 bool get hostIsWindows => false;
 bool get hostIsFuchsia => false;
+const Map<String, String> hostEnvironment = <String, String>{};
 String get hostOperatingSystem => 'web';
 String get hostOperatingSystemVersion => '';
 
@@ -22,3 +25,7 @@ Future<bool> hostCanResolve(String host,
 Future<String?> hostReadDocumentText(String name) async => null;
 
 Future<void> hostWriteDocumentText(String name, String contents) async {}
+
+/// The web file picker hands bytes back directly; there is no path to read.
+Future<Uint8List> hostReadFileBytes(String path) =>
+    throw UnsupportedError('hostReadFileBytes is not available on web');
