+++ lib/services/adblock_engine_io.dart
@@ -1,0 +1,745 @@
+// Dart FFI wrapper around the `webspace_adblock` Rust crate (which
+// itself wraps Brave's `adblock` engine).
+//
+// The Rust C FFI is in rust/webspace_adblock/src/lib.rs; the C
+// header is generated by cbindgen at rust/include/webspace_adblock.h.
+//
+// This wrapper is intentionally narrow: load the shared library,
+// expose a Dart-idiomatic [AdblockEngine] type, and convert between
+// Dart strings and the `(char*, size_t)` pairs the Rust side wants.
+//
+// Native artifacts must live where each platform's loader expects:
+//   * Android: `android/app/src/main/jniLibs/<abi>/libwebspace_adblock.so`
+//   * iOS / macOS: linked into the Runner app via XCFramework
+//     (`Adblock` symbol prefix is the same)
+//   * Linux: `linux/lib/libwebspace_adblock.so` packaged by the
+//     bundle plugin, or system-installed
+//   * Windows: not supported in the current spec
+//
+// On platforms that don't ship the library, [AdblockEngine.load]
+// returns null — callers MUST handle that case (the legacy Dart
+// content-blocker engine remains the fallback).
+
+import 'dart:convert';
+import 'dart:ffi' as ffi;
+import 'dart:io';
+import 'dart:typed_data';
+
+import 'package:ffi/ffi.dart';
+
+// ---------- C function signatures ----------
+
+typedef _EngineNewC = ffi.Pointer<ffi.Void> Function(
+    ffi.Pointer<Utf8>, ffi.UintPtr, ffi.Bool);
+typedef _EngineNewDart = ffi.Pointer<ffi.Void> Function(
+    ffi.Pointer<Utf8>, int, bool);
+
+typedef _EngineFreeC = ffi.Void Function(ffi.Pointer<ffi.Void>);
+typedef _EngineFreeDart = void Function(ffi.Pointer<ffi.Void>);
+
+typedef _EngineCheckUrlC = ffi.Int32 Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+);
+typedef _EngineCheckUrlDart = int Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+);
+
+typedef _EngineCosmeticC = ffi.Pointer<Utf8> Function(
+    ffi.Pointer<ffi.Void>, ffi.Pointer<Utf8>, ffi.UintPtr);
+typedef _EngineCosmeticDart = ffi.Pointer<Utf8> Function(
+    ffi.Pointer<ffi.Void>, ffi.Pointer<Utf8>, int);
+
+typedef _EngineHiddenClassIdC = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+);
+typedef _EngineHiddenClassIdDart = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+);
+
+typedef _FiltersToContentBlockingC = ffi.Pointer<Utf8> Function(
+    ffi.Pointer<Utf8>, ffi.UintPtr);
+typedef _FiltersToContentBlockingDart = ffi.Pointer<Utf8> Function(
+    ffi.Pointer<Utf8>, int);
+
+typedef _EngineRedirectForC = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+);
+typedef _EngineRedirectForDart = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+);
+
+typedef _EngineRewrittenUrlC = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+);
+typedef _EngineRewrittenUrlDart = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+);
+
+typedef _EngineCspForC = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+  ffi.Pointer<Utf8>,
+  ffi.UintPtr,
+);
+typedef _EngineCspForDart = ffi.Pointer<Utf8> Function(
+  ffi.Pointer<ffi.Void>,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+  ffi.Pointer<Utf8>,
+  int,
+);
+
+typedef _EngineSerializeC = ffi.Pointer<ffi.Uint8> Function(
+    ffi.Pointer<ffi.Void>, ffi.Pointer<ffi.UintPtr>);
+typedef _EngineSerializeDart = ffi.Pointer<ffi.Uint8> Function(
+    ffi.Pointer<ffi.Void>, ffi.Pointer<ffi.UintPtr>);
+
+typedef _EngineNewFromSerializedC = ffi.Pointer<ffi.Void> Function(
+    ffi.Pointer<ffi.Uint8>, ffi.UintPtr, ffi.Bool);
+typedef _EngineNewFromSerializedDart = ffi.Pointer<ffi.Void> Function(
+    ffi.Pointer<ffi.Uint8>, int, bool);
+
+typedef _BufFreeC = ffi.Void Function(ffi.Pointer<ffi.Uint8>, ffi.UintPtr);
+typedef _BufFreeDart = void Function(ffi.Pointer<ffi.Uint8>, int);
+
+typedef _StringFreeC = ffi.Void Function(ffi.Pointer<Utf8>);
+typedef _StringFreeDart = void Function(ffi.Pointer<Utf8>);
+
+typedef _EngineVersionC = ffi.Pointer<Utf8> Function();
+typedef _EngineVersionDart = ffi.Pointer<Utf8> Function();
+
+typedef _DepLicensesC = ffi.Pointer<Utf8> Function();
+typedef _DepLicensesDart = ffi.Pointer<Utf8> Function();
+
+/// Bundle of resolved C entry points for one library handle. Cached
+/// so the symbol-lookup cost is paid once per process, not per call.
+class _Bindings {
+  final _EngineNewDart engineNew;
+  final _EngineNewFromSerializedDart engineNewFromSerialized;
+  final _EngineFreeDart engineFree;
+  final _EngineCheckUrlDart engineCheckUrl;
+  final _EngineCosmeticDart engineCosmetic;
+  final _EngineHiddenClassIdDart engineHiddenClassId;
+  final _FiltersToContentBlockingDart filtersToContentBlocking;
+  final _EngineRedirectForDart engineRedirectFor;
+  final _EngineRewrittenUrlDart engineRewrittenUrl;
+  final _EngineCspForDart engineCspFor;
+  final _EngineSerializeDart engineSerialize;
+  final _BufFreeDart bufFree;
+  final _StringFreeDart stringFree;
+  final _EngineVersionDart engineVersion;
+  final _DepLicensesDart depLicenses;
+
+  _Bindings._({
+    required this.engineNew,
+    required this.engineNewFromSerialized,
+    required this.engineFree,
+    required this.engineCheckUrl,
+    required this.engineCosmetic,
+    required this.engineHiddenClassId,
+    required this.filtersToContentBlocking,
+    required this.engineRedirectFor,
+    required this.engineRewrittenUrl,
+    required this.engineCspFor,
+    required this.engineSerialize,
+    required this.bufFree,
+    required this.stringFree,
+    required this.engineVersion,
+    required this.depLicenses,
+  });
+
+  factory _Bindings.fromLib(ffi.DynamicLibrary lib) {
+    return _Bindings._(
+      engineNew: lib
+          .lookupFunction<_EngineNewC, _EngineNewDart>('ws_engine_new'),
+      engineNewFromSerialized: lib.lookupFunction<_EngineNewFromSerializedC,
+              _EngineNewFromSerializedDart>(
+          'ws_engine_new_from_serialized'),
+      engineFree: lib
+          .lookupFunction<_EngineFreeC, _EngineFreeDart>('ws_engine_free'),
+      engineCheckUrl: lib
+          .lookupFunction<_EngineCheckUrlC, _EngineCheckUrlDart>(
+              'ws_engine_check_url'),
+      engineCosmetic: lib.lookupFunction<_EngineCosmeticC, _EngineCosmeticDart>(
+          'ws_engine_cosmetic_resources_json'),
+      engineHiddenClassId: lib.lookupFunction<_EngineHiddenClassIdC,
+              _EngineHiddenClassIdDart>(
+          'ws_engine_hidden_class_id_selectors_json'),
+      filtersToContentBlocking: lib.lookupFunction<_FiltersToContentBlockingC,
+              _FiltersToContentBlockingDart>(
+          'ws_filters_to_content_blocking_json'),
+      engineRedirectFor: lib.lookupFunction<_EngineRedirectForC,
+              _EngineRedirectForDart>(
+          'ws_engine_redirect_for'),
+      engineRewrittenUrl: lib.lookupFunction<_EngineRewrittenUrlC,
+              _EngineRewrittenUrlDart>(
+          'ws_engine_rewritten_url'),
+      engineCspFor: lib
+          .lookupFunction<_EngineCspForC, _EngineCspForDart>(
+              'ws_engine_csp_for'),
+      engineSerialize: lib.lookupFunction<_EngineSerializeC,
+          _EngineSerializeDart>('ws_engine_serialize'),
+      bufFree:
+          lib.lookupFunction<_BufFreeC, _BufFreeDart>('ws_buf_free'),
+      stringFree:
+          lib.lookupFunction<_StringFreeC, _StringFreeDart>('ws_string_free'),
+      engineVersion: lib.lookupFunction<_EngineVersionC, _EngineVersionDart>(
+          'ws_engine_version'),
+      depLicenses: lib.lookupFunction<_DepLicensesC, _DepLicensesDart>(
+          'ws_dep_licenses_json'),
+    );
+  }
+}
+
+/// Engine instance. Owns native memory; call [dispose] to release.
+class AdblockEngine {
+  final _Bindings _b;
+  ffi.Pointer<ffi.Void> _handle;
+
+  AdblockEngine._(this._b, this._handle);
+
+  /// Try to load the shared library and parse [rulesText] into a
+  /// new engine. Returns `null` if the library can't be loaded on
+  /// this platform/arch, or if parsing fails. Caller is responsible
+  /// for logging null-return diagnostics — this layer is intentionally
+  /// dependency-free and doesn't pull in the app logger.
+  static AdblockEngine? load(String rulesText,
+      {bool enableUboResources = true}) {
+    final lib = _tryOpenLibrary();
+    if (lib == null) return null;
+    final bindings = _Bindings.fromLib(lib);
+    final bytes = utf8.encode(rulesText);
+    final ptr = malloc.allocate<ffi.Uint8>(bytes.length);
+    try {
+      ptr.asTypedList(bytes.length).setAll(0, bytes);
+      final handle = bindings.engineNew(
+          ptr.cast<Utf8>(), bytes.length, enableUboResources);
+      if (handle == ffi.nullptr) return null;
+      return AdblockEngine._(bindings, handle);
+    } finally {
+      malloc.free(ptr);
+    }
+  }
+
+  /// Build an engine from the binary blob produced by [serialize].
+/// Skips the multi-megabyte ABP rule parse — used to cache the
+/// parsed engine state across app launches.
+///
+/// Returns null when:
+///   * the library can't be loaded on this platform/arch,
+///   * the blob is corrupt / from an incompatible adblock-rust version,
+///   * a panic crosses the FFI boundary (release builds abort instead).
+///
+/// Caller should treat null as "cache miss, fall back to parse".
+  static AdblockEngine? loadFromSerialized(Uint8List bytes,
+      {bool enableUboResources = true}) {
+    if (bytes.isEmpty) return null;
+    final lib = _tryOpenLibrary();
+    if (lib == null) return null;
+    final bindings = _Bindings.fromLib(lib);
+    final ptr = malloc.allocate<ffi.Uint8>(bytes.length);
+    try {
+      ptr.asTypedList(bytes.length).setAll(0, bytes);
+      final handle = bindings.engineNewFromSerialized(
+          ptr, bytes.length, enableUboResources);
+      if (handle == ffi.nullptr) return null;
+      return AdblockEngine._(bindings, handle);
+    } finally {
+      malloc.free(ptr);
+    }
+  }
+
+  /// Serialize the engine's parsed state to a flatbuffer blob suitable
+  /// for [loadFromSerialized]. Used to cache the engine across app
+  /// launches and skip the multi-second rule parse on warm start.
+  ///
+  /// Returns null on any error (panic, OOM). Caller treats null as
+  /// "don't write a cache file this rebuild".
+  Uint8List? serialize() {
+    if (_handle == ffi.nullptr) return null;
+    final lenPtr = malloc.allocate<ffi.UintPtr>(1);
+    try {
+      final bufPtr = _b.engineSerialize(_handle, lenPtr);
+      if (bufPtr == ffi.nullptr) return null;
+      final len = lenPtr.value;
+      if (len == 0) {
+        _b.bufFree(bufPtr, 0);
+        return null;
+      }
+      // Copy out of native memory before freeing the buffer — the
+      // Uint8List we return must own a Dart heap allocation, not
+      // borrow from the C side.
+      final copy = Uint8List.fromList(bufPtr.asTypedList(len));
+      _b.bufFree(bufPtr, len);
+      return copy;
+    } finally {
+      malloc.free(lenPtr);
+    }
+  }
+
+  /// JSON-encoded list of the Rust-side transitive dependency tree
+  /// with SPDX licenses. Built at compile time by
+  /// `rust/webspace_adblock/build.rs` via `cargo metadata`. Surfaces
+  /// adblock-rust + every transitive crate's attribution without
+  /// vendoring any license text into the repo.
+  ///
+  /// Returns an empty list when the native library can't be loaded
+  /// on this platform — same fallback shape as the engine itself.
+  /// Static — no engine instance required.
+  static List<Map<String, dynamic>> depLicenses() {
+    final lib = _tryOpenLibrary();
+    if (lib == null) return const [];
+    final bindings = _Bindings.fromLib(lib);
+    final ptr = bindings.depLicenses();
+    if (ptr == ffi.nullptr) return const [];
+    // ws_dep_licenses_json returns a pointer to .rodata — DO NOT free.
+    final json = ptr.toDartString();
+    try {
+      final decoded = jsonDecode(json);
+      if (decoded is! List) return const [];
+      return decoded.cast<Map<String, dynamic>>();
+    } catch (_) {
+      return const [];
+    }
+  }
+
+  /// Engine library version string, for diagnostics.
+  String get version {
+    final ptr = _b.engineVersion();
+    if (ptr == ffi.nullptr) return '<unknown>';
+    final s = ptr.toDartString();
+    _b.stringFree(ptr);
+    return s;
+  }
+
+  /// Returns true if the engine would block [url] (when loaded as a
+  /// resource of [sourceUrl] / [requestType]).
+  ///
+  /// `requestType` follows ABP's resource-type taxonomy:
+  /// `document|subdocument|stylesheet|script|image|font|media|xhr|other`.
+  /// `other` is the safe default for "we don't know".
+  bool shouldBlock(
+    String url, {
+    String sourceUrl = '',
+    String requestType = 'other',
+  }) {
+    final urlBytes = utf8.encode(url);
+    final srcBytes = utf8.encode(sourceUrl);
+    final typBytes = utf8.encode(requestType);
+
+    final urlPtr = malloc.allocate<ffi.Uint8>(urlBytes.length);
+    final srcPtr = srcBytes.isEmpty
+        ? ffi.nullptr.cast<ffi.Uint8>()
+        : malloc.allocate<ffi.Uint8>(srcBytes.length);
+    final typPtr = malloc.allocate<ffi.Uint8>(typBytes.length);
+
+    try {
+      urlPtr.asTypedList(urlBytes.length).setAll(0, urlBytes);
+      if (srcBytes.isNotEmpty) {
+        srcPtr.asTypedList(srcBytes.length).setAll(0, srcBytes);
+      }
+      typPtr.asTypedList(typBytes.length).setAll(0, typBytes);
+
+      final result = _b.engineCheckUrl(
+        _handle,
+        urlPtr.cast<Utf8>(),
+        urlBytes.length,
+        srcPtr.cast<Utf8>(),
+        srcBytes.length,
+        typPtr.cast<Utf8>(),
+        typBytes.length,
+      );
+      return result == 1;
+    } finally {
+      malloc.free(urlPtr);
+      if (srcBytes.isNotEmpty) malloc.free(srcPtr);
+      malloc.free(typPtr);
+    }
+  }
+
+  /// Look up generic cosmetic selectors that target any of the given
+  /// [classes] / [ids]. Pair with [cosmeticResources] to get the full
+  /// per-page rule set: domain-scoped rules come from
+  /// [cosmeticResources], generic ones come from this method.
+  ///
+  /// Caller workflow:
+  ///   1. Page loads. JS scans the DOM for unique classes and ids.
+  ///   2. Send them through this method to the engine.
+  ///   3. Inject the returned selectors into a `<style>` tag with
+  ///      `display: none !important`.
+  ///
+  /// [exceptions] is the engine's `exceptions` list for the page,
+  /// already collected via [cosmeticResources]; pass an empty set if
+  /// not available — the worst case is a few false-positive hides.
+  ///
+  /// Returns an empty list on FFI error (instead of throwing) so the
+  /// caller can degrade gracefully.
+  List<String> hiddenClassIdSelectors(
+    Set<String> classes,
+    Set<String> ids, {
+    Set<String> exceptions = const <String>{},
+  }) {
+    final classesJson = jsonEncode(classes.toList());
+    final idsJson = jsonEncode(ids.toList());
+    final exceptionsJson = jsonEncode(exceptions.toList());
+
+    final classesBytes = utf8.encode(classesJson);
+    final idsBytes = utf8.encode(idsJson);
+    final excBytes = utf8.encode(exceptionsJson);
+
+    final classesPtr = malloc.allocate<ffi.Uint8>(classesBytes.length);
+    final idsPtr = malloc.allocate<ffi.Uint8>(idsBytes.length);
+    final excPtr = malloc.allocate<ffi.Uint8>(excBytes.length);
+
+    try {
+      classesPtr.asTypedList(classesBytes.length).setAll(0, classesBytes);
+      idsPtr.asTypedList(idsBytes.length).setAll(0, idsBytes);
+      excPtr.asTypedList(excBytes.length).setAll(0, excBytes);
+
+      final ptr = _b.engineHiddenClassId(
+        _handle,
+        classesPtr.cast<Utf8>(),
+        classesBytes.length,
+        idsPtr.cast<Utf8>(),
+        idsBytes.length,
+        excPtr.cast<Utf8>(),
+        excBytes.length,
+      );
+      if (ptr == ffi.nullptr) return const [];
+      final json = ptr.toDartString();
+      _b.stringFree(ptr);
+      final decoded = jsonDecode(json);
+      if (decoded is! List) return const [];
+      return decoded.cast<String>();
+    } finally {
+      malloc.free(classesPtr);
+      malloc.free(idsPtr);
+      malloc.free(excPtr);
+    }
+  }
+
+  /// If a `$redirect=` (or `$redirect-rule=`) rule matches this
+  /// request, returns the redirect target as a data: URL
+  /// (`data:<mime>;base64,<body>`). Returns null when no redirect
+  /// applies — caller should fall back to the empty-body block
+  /// response.
+  ///
+  /// Called from the iOS/macOS JS bridge's `blockCheck` handler so
+  /// the JS interceptor can swap the request URL with the data URL
+  /// instead of dropping it. Same FFI symbol the JNI bridge calls
+  /// on Android (`ws_engine_redirect_for`).
+  String? redirectFor(
+    String url, {
+    String sourceUrl = '',
+    String requestType = 'other',
+  }) {
+    final urlBytes = utf8.encode(url);
+    final srcBytes = utf8.encode(sourceUrl);
+    final typBytes = utf8.encode(requestType);
+
+    final urlPtr = malloc.allocate<ffi.Uint8>(urlBytes.length);
+    final srcPtr = srcBytes.isEmpty
+        ? ffi.nullptr.cast<ffi.Uint8>()
+        : malloc.allocate<ffi.Uint8>(srcBytes.length);
+    final typPtr = malloc.allocate<ffi.Uint8>(typBytes.length);
+
+    try {
+      urlPtr.asTypedList(urlBytes.length).setAll(0, urlBytes);
+      if (srcBytes.isNotEmpty) {
+        srcPtr.asTypedList(srcBytes.length).setAll(0, srcBytes);
+      }
+      typPtr.asTypedList(typBytes.length).setAll(0, typBytes);
+
+      final resultPtr = _b.engineRedirectFor(
+        _handle,
+        urlPtr.cast<Utf8>(),
+        urlBytes.length,
+        srcPtr.cast<Utf8>(),
+        srcBytes.length,
+        typPtr.cast<Utf8>(),
+        typBytes.length,
+      );
+      if (resultPtr == ffi.nullptr) return null;
+      final dataUrl = resultPtr.toDartString();
+      _b.stringFree(resultPtr);
+      return dataUrl;
+    } finally {
+      malloc.free(urlPtr);
+      if (srcBytes.isNotEmpty) malloc.free(srcPtr);
+      malloc.free(typPtr);
+    }
+  }
+
+  /// `$removeparam=` lookup. Returns the URL after adblock-rust strips
+  /// matching query parameters. Returns null when no rewrite applies.
+  ///
+  /// This is independent of [shouldBlock]: a URL can be both blocked
+  /// AND rewritten. Callers typically check block first; if not
+  /// blocked, check rewritten then issue the rewritten URL.
+  String? rewrittenUrl(
+    String url, {
+    String sourceUrl = '',
+    String requestType = 'other',
+  }) =>
+      _callStringFn(
+        _b.engineRewrittenUrl,
+        url,
+        sourceUrl,
+        requestType,
+      );
+
+  /// `$csp=` lookup. Returns the CSP directives the engine wants to
+  /// apply at this URL (joined `;`-separated string). Returns null
+  /// when no `$csp=` rule matches.
+  ///
+  /// Caller injects the result either as a response header
+  /// (Android, where we control sub-resource responses) or as a
+  /// `<meta http-equiv="Content-Security-Policy">` tag injected at
+  /// DOCUMENT_START (Apple, where we can't rewrite headers).
+  String? cspFor(
+    String url, {
+    String sourceUrl = '',
+    String requestType = 'other',
+  }) =>
+      _callStringFn(
+        _b.engineCspFor,
+        url,
+        sourceUrl,
+        requestType,
+      );
+
+  /// Helper to invoke the (engine, url, src, type) → cstring FFI shape
+  /// without duplicating the malloc/copy/free boilerplate.
+  String? _callStringFn(
+    ffi.Pointer<Utf8> Function(
+      ffi.Pointer<ffi.Void>,
+      ffi.Pointer<Utf8>,
+      int,
+      ffi.Pointer<Utf8>,
+      int,
+      ffi.Pointer<Utf8>,
+      int,
+    ) fn,
+    String url,
+    String sourceUrl,
+    String requestType,
+  ) {
+    final urlBytes = utf8.encode(url);
+    final srcBytes = utf8.encode(sourceUrl);
+    final typBytes = utf8.encode(requestType);
+
+    final urlPtr = malloc.allocate<ffi.Uint8>(urlBytes.length);
+    final srcPtr = srcBytes.isEmpty
+        ? ffi.nullptr.cast<ffi.Uint8>()
+        : malloc.allocate<ffi.Uint8>(srcBytes.length);
+    final typPtr = malloc.allocate<ffi.Uint8>(typBytes.length);
+
+    try {
+      urlPtr.asTypedList(urlBytes.length).setAll(0, urlBytes);
+      if (srcBytes.isNotEmpty) {
+        srcPtr.asTypedList(srcBytes.length).setAll(0, srcBytes);
+      }
+      typPtr.asTypedList(typBytes.length).setAll(0, typBytes);
+
+      final ptr = fn(
+        _handle,
+        urlPtr.cast<Utf8>(),
+        urlBytes.length,
+        srcPtr.cast<Utf8>(),
+        srcBytes.length,
+        typPtr.cast<Utf8>(),
+        typBytes.length,
+      );
+      if (ptr == ffi.nullptr) return null;
+      final s = ptr.toDartString();
+      _b.stringFree(ptr);
+      return s;
+    } finally {
+      malloc.free(urlPtr);
+      if (srcBytes.isNotEmpty) malloc.free(srcPtr);
+      malloc.free(typPtr);
+    }
+  }
+
+  /// Per-URL cosmetic resources from the engine, decoded from JSON.
+  /// Returns null if the engine reports nothing or hits an error.
+  ///
+  /// The shape mirrors `adblock::Engine::url_cosmetic_resources`:
+  /// hide selectors (domain-specific only — generic class/id rules
+  /// need a separate scan-and-query pass that's not wired up yet),
+  /// procedural actions, exceptions, injected_script, generichide.
+  Map<String, dynamic>? cosmeticResources(String url) {
+    final urlBytes = utf8.encode(url);
+    final urlPtr = malloc.allocate<ffi.Uint8>(urlBytes.length);
+    try {
+      urlPtr.asTypedList(urlBytes.length).setAll(0, urlBytes);
+      final ptr = _b.engineCosmetic(_handle, urlPtr.cast<Utf8>(), urlBytes.length);
+      if (ptr == ffi.nullptr) return null;
+      final json = ptr.toDartString();
+      _b.stringFree(ptr);
+      return jsonDecode(json) as Map<String, dynamic>;
+    } finally {
+      malloc.free(urlPtr);
+    }
+  }
+
+  /// Convert an ABP filter list to Apple's WKContentRuleListStore
+  /// JSON format. Static — needs no engine instance (this is a
+  /// pure parse+convert pass through the underlying crate).
+  ///
+  /// The result is suitable for
+  /// `WKContentRuleListStore.compileContentRuleList`. Returns null
+  /// when the native library isn't loadable on this platform, or
+  /// when the filter list fails to parse.
+  ///
+  /// Heads-up: WebKit doesn't fire any callback when a content rule
+  /// matches. Use this for the speedup (native sub-resource blocking
+  /// without a JS bridge round-trip per request), but keep the JS
+  /// bridge installed if you want per-request stats.
+  static String? filterListToAppleContentBlockingJson(String rulesText) {
+    final lib = _tryOpenLibrary();
+    if (lib == null) return null;
+    final bindings = _Bindings.fromLib(lib);
+    final bytes = utf8.encode(rulesText);
+    final ptr = malloc.allocate<ffi.Uint8>(bytes.length);
+    try {
+      ptr.asTypedList(bytes.length).setAll(0, bytes);
+      final resultPtr = bindings.filtersToContentBlocking(
+        ptr.cast<Utf8>(),
+        bytes.length,
+      );
+      if (resultPtr == ffi.nullptr) return null;
+      final json = resultPtr.toDartString();
+      bindings.stringFree(resultPtr);
+      return json;
+    } finally {
+      malloc.free(ptr);
+    }
+  }
+
+  /// Release the engine. Safe to call multiple times.
+  void dispose() {
+    if (_handle == ffi.nullptr) return;
+    _b.engineFree(_handle);
+    _handle = ffi.nullptr;
+  }
+}
+
+/// Resolve the native library on the current platform. Returns null
+/// when the library isn't shipped — callers must fall back to the
+/// legacy Dart engine in that case.
+ffi.DynamicLibrary? _tryOpenLibrary() {
+  // Test runs (`flutter test`, `dart test`) execute on the host
+  // machine, NOT inside the bundled app, so the per-platform paths
+  // below don't apply. Look for the cargo `target/release/` artifact
+  // first — that's what `cargo build --release` in
+  // rust/webspace_adblock leaves behind.
+  if (Platform.isLinux || Platform.isMacOS) {
+    for (final candidate in _testCandidatePaths()) {
+      try {
+        return ffi.DynamicLibrary.open(candidate);
+      } catch (_) {
+        continue;
+      }
+    }
+  }
+
+  try {
+    if (Platform.isAndroid) {
+      return ffi.DynamicLibrary.open('libwebspace_adblock.so');
+    }
+    if (Platform.isLinux) {
+      // Bundled installs land the .so at `<bundle>/lib/libwebspace_
+      // adblock.so` next to the Flutter engine library; the binary
+      // is built with RPATH=$ORIGIN/lib so a bare basename resolves.
+      // Fall through to absolute path if the bare lookup fails (e.g.
+      // running unbundled from `flutter run`).
+      try {
+        return ffi.DynamicLibrary.open('libwebspace_adblock.so');
+      } catch (_) {
+        final exeDir = File(Platform.resolvedExecutable).parent.path;
+        return ffi.DynamicLibrary.open('$exeDir/lib/libwebspace_adblock.so');
+      }
+    }
+    if (Platform.isMacOS || Platform.isIOS) {
+      // Both Apple targets ship the rust .a inside a CocoaPods-built
+      // framework — webspace_adblock.framework — embedded in the
+      // Runner.app bundle (see rust/webspace_adblock/cocoapods/). Its
+      // symbols are loaded into the process at startup via the
+      // framework's dyld dependency, so DynamicLibrary.process() finds
+      // them through dlsym(RTLD_DEFAULT, ...). keep_alive.c in the pod
+      // takes their addresses at compile time to defeat dead_strip.
+      return ffi.DynamicLibrary.process();
+    }
+  } catch (_) {
+    return null;
+  }
+  return null;
+}
+
+List<String> _testCandidatePaths() {
+  // Two candidates so the test harness works regardless of CWD:
+  // running `flutter test` from the repo root lands at
+  // rust/webspace_adblock/target/...; running it from another tree
+  // (rare) we still try the absolute path computed off Platform
+  // resolved executable. Keep the list small to avoid stat thrash.
+  final cwd = Directory.current.path;
+  final ext = Platform.isMacOS ? 'dylib' : 'so';
+  return <String>[
+    '$cwd/rust/webspace_adblock/target/release/libwebspace_adblock.$ext',
+    '$cwd/rust/webspace_adblock/target/debug/libwebspace_adblock.$ext',
+  ];
+}
